backend/requirements.txt
fastapi>=0.95.0,<0.105.0
uvicorn>=0.22.0,<0.24.0
httpx>=0.24.0,<0.26.0
beautifulsoup4>=4.11.0,<4.13.0
requests>=2.28.0,<2.33.0
python-dotenv>=0.21.0,<1.1.0
google-auth>=2.16.0,<2.24.0
google-auth-oauthlib>=1.0.0,<1.2.0
google-api-python-client>=2.80.0,<2.110.0
pydantic>=2.0.0,<2.5.0
python-multipart>=0.0.5,<0.0.7
itsdangerous>=2.0.0,<3.0.0
lxml_html_clean
pydantic[email]